#620: New core plugin
