+++ lib/cli.js
@@ -99,6 +99,8 @@ exports.largv = this.parseGlobals();
  *
  * The interactivity metadata is light wrapper around [inquirer](https://github.com/sboudrias/Inquirer.js)
  *
+ * @TODO: Injecting the events here seems not the way we want to go?
+ *
  * @since 3.0.0
  * @see [yargs docs](http://yargs.js.org/docs/)
  * @see [inquirer docs](https://github.com/sboudrias/Inquirer.js)
@@ -109,7 +111,7 @@ exports.largv = this.parseGlobals();
  * // Add that task to the CLI
  * yargs.command(lando.tasks.parseToYargs(task));
  */
-exports.parseToYargs = function(task) {
+exports.parseToYargs = function(task, events) {
 
   // Start command and option collectors
   var command = {};
@@ -135,8 +137,7 @@ exports.parseToYargs = function(task) {
   // Build a handler that incorporates interactive options
   command.handler = function(argv) {
 
-    // Get lando
-    var lando = require('./lando')(require('./config'));
+    // Build commands
     var cmd = argv._[0];
     var answersEvent = ['task', cmd, 'answers'].join('-');
     var runEvent = ['task', cmd, 'run'].join('-');
@@ -149,7 +150,7 @@ exports.parseToYargs = function(task) {
      * @event module:tasks.event:task-CMD-answers
      * @property {Object} answers argv and inquirer questions
      */
-    return lando.events.emit(answersEvent, {argv: argv, inquirer: inquiry})
+    return events.emit(answersEvent, {argv: argv, inquirer: inquiry})
 
     // Do the filtering
     .then(function() {
@@ -177,7 +178,7 @@ exports.parseToYargs = function(task) {
        * @event module:tasks.event:task-CMD-run
        * @property {Object} answers object
        */
-      return lando.events.emit(runEvent, truth)
+      return events.emit(runEvent, truth)
 
       // Run the task
       .then(function() {
