+++ plugins/lando-core/tasks/version.js
